// libs/shared/exchange-rate/data-access/src/lib/models/exhange-rates.model.ts
import {CurrencyModel} from "@monorepo-tools/shared/exchange-rate/util";

export type ExchangeRateModel = {
  result: string;
  documentation: string;
  terms_of_use: string;
  time_last_update_unix: number;
  "time_last_update_utc": string;
  time_next_update_unix: number;
  time_next_update_utc: string;
  base_code: CurrencyModel;
  target_code: CurrencyModel;
  "conversion_rate": number;
}